Add initial version of a protobuf style definition for the storage API
This is my proposal for the minimum API surface that would be usable by
both clients writing data to the backend, and consumers reading
encrypted data for downstream use. It does not include any API
affordances for deleting or modifying data. It does not provide any way
of requesting data other than by a user's public key, with optional time
based filtering.